+++ angular-project/projects/otec/src/app/features/cubicacion/crear-cubicacion/crear-cubicacion.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'otec-crear-cubicacion',
+  templateUrl: './crear-cubicacion.component.html',
+  styleUrls: ['./crear-cubicacion.component.css']
+})
+export class CrearCubicacionComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}