+++ angular-project/projects/otec/src/app/shared/shared.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {NavbarComponent} from '../features/navbar/navbar.component'
+import { NavbarComponent } from '../features/navbar/navbar.component';
 @NgModule({
   declarations: [NavbarComponent],
   imports: [CommonModule],
-  exports: [CommonModule,NavbarComponent],
+  exports: [CommonModule, NavbarComponent],
 })
 export class SharedModule {}